backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
yfinance
httpx
pandas
